refactor(sync): move sync value types to Domain, drop singleton access in view
- Extract SyncTrigger + SyncMetadata to Domain/Models/SyncMetadata.swift
- Remove them from CloudSaveSyncSettings (Data layer)
- Route lastSync(romId:) through LocalDeviceDetailViewModel instead of direct CloudSaveSyncSettings.shared access in PlatformROMsListView
- Consolidate RelativeDateTimeFormatter into Date+RelativeFormatting.swift extension; remove duplicate static func from SyncSaveSheet and inline formatter from SyncStatusLine
- recordManualSync() sets lastSyncMeta directly without a UserDefaults round-trip
